Convert the solver to the async world
`Solver` and `Problem` are now completely async. Despite this, the
solver algorithm itself is not parallel. The intent of this change is to
allow `Problem` implementations to be async, and the solver will not
lock its internal state while `Problem::direct_dependencies` or
`Problem::evaluate` are running, allowing those methods to run
concurrently.

As a side effect, the solver should run slightly slower even in a
single-threaded scenario. An configurable async/blocking solver was
deemed too complex. However it is possible to implement dummy mutexes in
the future to recover most if not all of the lost performance in a
single-threaded scenario.

`Solver::clone_with_evaluation_assumptions` was temporarely removed, as
were some trait implementations that need to be rethought in an async
world.

Next step to solving the cloning and parallelization problems would be
to use lock-free structures with some extra synchronization to emulate
the small transactions that the solver performs.

This commit also includes some QoL changes.